refactor(website): Extract HTTP response utilities to dedicated http.ts
- Move createErrorResponse function and ErrorResponse interface from logger.ts to utils/http.ts
- Update imports across packAction.ts, rateLimit.ts, and bodyLimit.ts
- Improve separation of concerns: logging utilities vs HTTP response utilities
- Maintains consistent error response format across the application
