+++ website/server/src/actions/packAction.ts
@@ -7,7 +7,8 @@ import { FILE_SIZE_LIMITS } from '../domains/pack/utils/fileUtils.js';
 import { sanitizePattern } from '../domains/pack/utils/validation.js';
 import type { PackResult } from '../types.js';
 import { getClientInfo } from '../utils/clientInfo.js';
-import { createErrorResponse, logError, logInfo } from '../utils/logger.js';
+import { createErrorResponse } from '../utils/http.js';
+import { logError, logInfo } from '../utils/logger.js';
 import { calculateMemoryDiff, getMemoryUsage } from '../utils/memory.js';
 import { formatLatencyForDisplay } from '../utils/time.js';
 import { validateRequest } from '../utils/validation.js';
